修正多模值，解决编译错误

+++ window_scene/session_manager/src/scene_session_manager.cpp
@@ -11866,7 +11866,7 @@ void SceneSessionManager::NotifyAppUseControlListInner(
     std::vector<sptr<SceneSession>> mainSessions;
     std::vector<AppUseControlInfo> controlByBundleList;
     for (const auto& appUseControlInfo : controlList) {
-        if (appUseControlInfo.persistentId_ > static_cast<int32_t>(INVALID_WINDOW_ID)) {
+        if (appUseControlInfo.persistentId_ > INVALID_WINDOW_ID) {
             // control by peristentId
             TLOGI(WmsLogTag::WMS_MAIN, "control by id:%{public}d", appUseControlInfo.persistentId_);
             auto iter = sceneSessionMap_.find(appUseControlInfo.persistentId_);
